requirements.txt
python-dotenv==1.0.0
APScheduler==3.10.4
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
executing==2.0.1
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pip==23.2.1
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
tzlocal==5.2
wcwidth==0.2.10